header and footer completed
