+++ app/javascript/QuizHeader.vue
@@ -1,5 +1,11 @@
 <template>
-  <p>{{ message }}</p>
+  <nav class="navbar navbar-expand-lg navbar-light header" style="background-color: #e3f2fd;">
+    <router-link :to="{ name: 'Index'}" id="router-link">
+      <span class="nav-title">
+        いや〜クイズ -30sの思い出-
+      </span>
+    </router-link>
+  </nav>
 </template>
 
 <script>
