+++ app/javascript/QuizPage1.vue
@@ -4,7 +4,7 @@
       <quiz-header></quiz-header>
     </header>
 
-    <main>
+    <main class="container">
       <quiz-content></quiz-content>
     </main>
 
@@ -18,13 +18,12 @@
 import QuizHeader from "QuizHeader.vue"
 import QuizFooter from "QuizFooter.vue"
 import QuizContent from "QuizContent.vue"
-import QuizResult from "QuizResult.vue"
+
 import axios from 'axios';
 
 export default {
   name: "QuizIndex",
   components: {
-    QuizResult,
     QuizHeader,
     QuizFooter,
     QuizContent
