+++ app/javascript/app.vue
@@ -6,12 +6,17 @@
   import Vue from 'vue'
   import VueRouter from 'vue-router'
 
-  import QuizIndex from 'QuizIndex.vue'
+  import Index from 'Index.vue'
+  import QuizPage1 from 'QuizPage1.vue'
 
   const router = new VueRouter ({
     routes: [
       { path: '/',
-        component: QuizIndex }
+        name: 'Index',
+        component: Index },
+      { path: '/quiz/page1/',
+        name: 'QuizPage1',
+        component: QuizPage1}
     ]
   })
   Vue.use(VueRouter)
